Changelog:
Part 2 of porting, it runs! Need to fix Azurelib armor crashing cause of pTag being null

+++ src/main/java/net/acetheeldritchking/cataclysm_spellbooks/spells/ender/VoidRuneBulwarkSpell.java
@@ -72,7 +72,7 @@ public class VoidRuneBulwarkSpell extends AbstractSpell {
 
     private void spawnVoidRune(Level level, double x, double z, double minY, double maxY, double rotation, int delay, LivingEntity caster, int spellLevel)
     {
-        BlockPos blockPos = new BlockPos(x, maxY, z);
+        BlockPos blockPos = new BlockPos((int) x, (int) maxY, (int) z);
         boolean flag = false;
         double d0 = 0.0D;
